♻️ [refactor]: #452 Phase 8-F: lexer_read_integer_then_read_real.rs を分離
read_integer の '.' 遭遇時の巻き戻し 4 テスト（Phase 8-F）を
`src/lexer/tests/lexer_read_integer_then_read_real.rs` に分離。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_read_integer_then_read_real.rs: 35 行 (< 800)

Refs #452
